Create 2017-10-01-How I came to join the Cooperative.md
Spencer, I just wrote this, I don't know if it fits and it is very poorly written. I would love if you would please review it, change it as much as you see fit, and do whatever you think its best with it, I really don't mind even if it gets reduced to a single paraghraph. I was writing it without any self-judgment at all, I just wrote what I thought like I would be speaking. If also you don't want to review it and change it yourself you can tell me and I can do that at a later time. My main concern is if it fits to write this story as a blog entry or if it even fits at all. Please give me some feedback.