Update transactions.yml
I want to apologize for not noting the exact times. I will focus on this more in the future. Also I would like to know if this is okay like this. Specially connected to the grocery shopping I want to know if it is disired to have more detailed information. I kept the receipt so it is possible. Mucho Amore